Create README.md
🌟 SuperHero🦸App 🌟:You can search your favorite SuperHero and Know their powerstats && and also generate Random SuperHeros.

In this project I used "superhero.api" API which contains data and information about a list of superheroes from the comic universe. This includes their physical capabilities, biography information, occupations, power statistics, connections with other heroes, and appearances in movies.
and used concepts like:
👉Fetch()
👉Template Literals
👉Higher Order Functions
👉Event Listeners
👉Dom Manipulation .etc

It is very organised and easy to Understand.